refactor(server): share strict HTTP date parsing
